Separate serialize structs and methods
