+++ cx/ast/serialize.go
@@ -6,215 +6,7 @@ import (
 	"github.com/skycoin/skycoin/src/cipher/encoder"
 )
 
-type serializedCXProgramIndex struct {
-	ProgramOffset     int64
-	CallsOffset       int64
-	PackagesOffset    int64
-	StructsOffset     int64
-	FunctionsOffset   int64
-	ExpressionsOffset int64
-	ArgumentsOffset   int64
-	IntegersOffset    int64
-	StringsOffset     int64
-	MemoryOffset      int64
-}
-
-type serializedProgram struct {
-	PackagesOffset       int64
-	PackagesSize         int64
-	CurrentPackageOffset int64
-
-	InputsOffset int64
-	InputsSize   int64
-
-	OutputsOffset int64
-	OutputsSize   int64
-
-	CallStackOffset int64
-	CallStackSize   int64
-
-	CallCounter int64
-
-	MemoryOffset int64
-	MemorySize   int64
-
-	StackPointer int64
-	StackSize    int64
-
-	DataSegmentSize     int64
-	DataSegmentStartsAt int64
-
-	HeapPointer  int64 //HeapPointer is probably related to HeapStartsAt
-	HeapStartsAt int64
-	HeapSize     int64
-
-	Terminated int64
-
-	VersionOffset int64
-	VersionSize   int64
-}
-
-type serializedCall struct {
-	OperatorOffset int64
-	Line           int64
-	FramePointer   int64
-}
-
-type serializedPackage struct {
-	NameOffset            int64
-	NameSize              int64
-	ImportsOffset         int64
-	ImportsSize           int64
-	StructsOffset         int64
-	StructsSize           int64
-	GlobalsOffset         int64
-	GlobalsSize           int64
-	FunctionsOffset       int64
-	FunctionsSize         int64
-	CurrentFunctionOffset int64
-	CurrentStructOffset   int64
-}
-
-type serializedStruct struct {
-	NameOffset   int64
-	NameSize     int64
-	FieldsOffset int64
-	FieldsSize   int64
-
-	Size int64
-
-	PackageOffset int64
-}
-
-type serializedFunction struct {
-	NameOffset        int64
-	NameSize          int64
-	InputsOffset      int64
-	InputsSize        int64
-	OutputsOffset     int64
-	OutputsSize       int64
-	ExpressionsOffset int64
-	ExpressionsSize   int64
-	Size              int64
-	Length            int64
-
-	ListOfPointersOffset int64
-	ListOfPointersSize   int64
-
-	// We're going to determine this when procesing the expressions. Check serializedExpression type
-	// IsBuiltin                        int64
-	// OpCode                          int64
-
-	CurrentExpressionOffset int64
-	PackageOffset           int64
-}
-
-type serializedExpression struct {
-	OperatorOffset int64
-	// we add these two fields here so we don't add every native serializedFunction to the serialization
-	// the CX runtime already knows about the natives properties. We just need the code if IsNative = true
-	IsNative int64
-	OpCode   int64
-
-	InputsOffset  int64
-	InputsSize    int64
-	OutputsOffset int64
-	OutputsSize   int64
-
-	LabelOffset int64
-	LabelSize   int64
-	ThenLines   int64
-	ElseLines   int64
-
-	ExpressionType int64
-
-	FunctionOffset int64
-	PackageOffset  int64
-}
-
-type serializedArgument struct {
-	NameOffset       int64
-	NameSize         int64
-	Type             int64
-	StructTypeOffset int64 //WTF IS A CUSTOM TYPE!?
-	Size             int64
-	TotalSize        int64
-
-	Offset int64
-
-	IndirectionLevels           int64
-	DereferenceLevels           int64
-	DereferenceOperationsOffset int64
-	DereferenceOperationsSize   int64
-	DeclarationSpecifiersOffset int64
-	DeclarationSpecifiersSize   int64
-
-	IsSlice      int64
-	IsArray      int64
-	IsArrayFirst int64
-	IsPointer    int64
-	IsReference  int64
-
-	IsDereferenceFirst int64
-	IsStruct           int64
-	IsRest             int64
-	IsLocalDeclaration int64
-	IsShortDeclaration int64
-	PreviouslyDeclared int64
-
-	PassBy     int64
-	DoesEscape int64
-
-	LengthsOffset int64
-	LengthsSize   int64
-	IndexesOffset int64
-	IndexesSize   int64
-	FieldsOffset  int64
-	FieldsSize    int64
-	InputsOffset  int64
-	InputsSize    int64
-	OutputsOffset int64
-	OutputsSize   int64
-
-	PackageOffset int64
-}
-
-type serializedCXProgram struct {
-	Index   serializedCXProgramIndex
-	Program serializedProgram
-
-	Packages     []serializedPackage
-	PackagesMap  map[string]int
-	Structs      []serializedStruct
-	StructsMap   map[string]int
-	Functions    []serializedFunction
-	FunctionsMap map[string]int
-
-	Expressions []serializedExpression
-	Arguments   []serializedArgument
-	Calls       []serializedCall
-
-	Strings    []byte
-	StringsMap map[string]int
-	Integers   []int64
-
-	Memory []byte
-}
-
-type SerializedDataSize struct {
-	Program     int `json:"program"`
-	Calls       int `json:"calls"`
-	Packages    int `json:"packages"`
-	Structs     int `json:"structs"`
-	Functions   int `json:"functions"`
-	Expressions int `json:"expressions"`
-	Arguments   int `json:"arguments"`
-	Integers    int `json:"integers"`
-	Strings     int `json:"strings"`
-	Memory      int `json:"memory"`
-}
-
-func serializeString(name string, s *serializedCXProgram) (int64, int64) {
+func serializeString(name string, s *SerializedCXProgram) (int64, int64) {
 	if name == "" {
 		return int64(-1), int64(-1)
 	}
@@ -225,34 +17,34 @@ func serializeString(name string, s *serializedCXProgram) (int64, int64) {
 	if found {
 		return int64(off), int64(size)
 	}
-	off = len(s.Strings)
+	off = int64(len(s.Strings))
 	s.Strings = append(s.Strings, encoder.Serialize(name)...)
 	s.StringsMap[name] = off
 
 	return int64(off), int64(size)
 }
 
-func indexPackage(pkg *CXPackage, s *serializedCXProgram) {
+func indexPackage(pkg *CXPackage, s *SerializedCXProgram) {
 	if _, found := s.PackagesMap[pkg.Name]; !found {
-		s.PackagesMap[pkg.Name] = len(s.PackagesMap)
+		s.PackagesMap[pkg.Name] = int64(len(s.PackagesMap))
 	} else {
 		panic("duplicated package in serialization process")
 	}
 }
 
-func indexStruct(strct *CXStruct, s *serializedCXProgram) {
+func indexStruct(strct *CXStruct, s *SerializedCXProgram) {
 	strctName := strct.Package.Name + "." + strct.Name
 	if _, found := s.StructsMap[strctName]; !found {
-		s.StructsMap[strctName] = len(s.StructsMap)
+		s.StructsMap[strctName] = int64(len(s.StructsMap))
 	} else {
 		panic("duplicated struct in serialization process")
 	}
 }
 
-func indexFunction(fn *CXFunction, s *serializedCXProgram) {
+func indexFunction(fn *CXFunction, s *SerializedCXProgram) {
 	fnName := fn.Package.Name + "." + fn.Name
 	if _, found := s.FunctionsMap[fnName]; !found {
-		s.FunctionsMap[fnName] = len(s.FunctionsMap)
+		s.FunctionsMap[fnName] = int64(len(s.FunctionsMap))
 	} else {
 		panic("duplicated function in serialization process")
 	}
@@ -265,7 +57,7 @@ func serializeBoolean(val bool) int64 {
 	return 0
 }
 
-func serializeIntegers(ints []int, s *serializedCXProgram) (int64, int64) {
+func serializeIntegers(ints []int, s *SerializedCXProgram) (int64, int64) {
 	if len(ints) == 0 {
 		return int64(-1), int64(-1)
 	}
@@ -282,7 +74,7 @@ func serializeIntegers(ints []int, s *serializedCXProgram) (int64, int64) {
 	return int64(off), int64(l)
 }
 
-func serializeArgument(arg *CXArgument, s *serializedCXProgram) int {
+func serializeArgument(arg *CXArgument, s *SerializedCXProgram) int {
 	s.Arguments = append(s.Arguments, serializedArgument{})
 	argOff := len(s.Arguments) - 1
 
@@ -346,7 +138,7 @@ func serializeArgument(arg *CXArgument, s *serializedCXProgram) int {
 	return argOff
 }
 
-func serializeSliceOfArguments(args []*CXArgument, s *serializedCXProgram) (int64, int64) {
+func serializeSliceOfArguments(args []*CXArgument, s *SerializedCXProgram) (int64, int64) {
 	if len(args) == 0 {
 		return int64(-1), int64(-1)
 	}
@@ -357,7 +149,7 @@ func serializeSliceOfArguments(args []*CXArgument, s *serializedCXProgram) (int6
 	return serializeIntegers(idxs, s)
 }
 
-func serializeCalls(calls []CXCall, s *serializedCXProgram) (int64, int64) {
+func serializeCalls(calls []CXCall, s *SerializedCXProgram) (int64, int64) {
 	if len(calls) == 0 {
 		return int64(-1), int64(-1)
 	}
@@ -369,7 +161,7 @@ func serializeCalls(calls []CXCall, s *serializedCXProgram) (int64, int64) {
 
 }
 
-func serializeExpression(expr *CXExpression, s *serializedCXProgram) int {
+func serializeExpression(expr *CXExpression, s *SerializedCXProgram) int {
 	s.Expressions = append(s.Expressions, serializedExpression{})
 	exprOff := len(s.Expressions) - 1
 	sExpr := &s.Expressions[exprOff]
@@ -420,7 +212,7 @@ func serializeExpression(expr *CXExpression, s *serializedCXProgram) int {
 	return exprOff
 }
 
-func serializeCall(call *CXCall, s *serializedCXProgram) int {
+func serializeCall(call *CXCall, s *SerializedCXProgram) int {
 	s.Calls = append(s.Calls, serializedCall{})
 	callOff := len(s.Calls) - 1
 	serializedCall := &s.Calls[callOff]
@@ -437,7 +229,7 @@ func serializeCall(call *CXCall, s *serializedCXProgram) int {
 	return callOff
 }
 
-func serializeStructArguments(strct *CXStruct, s *serializedCXProgram) {
+func serializeStructArguments(strct *CXStruct, s *SerializedCXProgram) {
 	strctName := strct.Package.Name + "." + strct.Name
 	if strctOff, found := s.StructsMap[strctName]; found {
 		sStrct := &s.Structs[strctOff]
@@ -447,7 +239,7 @@ func serializeStructArguments(strct *CXStruct, s *serializedCXProgram) {
 	}
 }
 
-func serializeFunctionArguments(fn *CXFunction, s *serializedCXProgram) {
+func serializeFunctionArguments(fn *CXFunction, s *SerializedCXProgram) {
 	fnName := fn.Package.Name + "." + fn.Name
 	if fnOff, found := s.FunctionsMap[fnName]; found {
 		sFn := &s.Functions[fnOff]
@@ -460,18 +252,18 @@ func serializeFunctionArguments(fn *CXFunction, s *serializedCXProgram) {
 	}
 }
 
-func serializePackageName(pkg *CXPackage, s *serializedCXProgram) {
+func serializePackageName(pkg *CXPackage, s *SerializedCXProgram) {
 	sPkg := &s.Packages[s.PackagesMap[pkg.Name]]
 	sPkg.NameOffset, sPkg.NameSize = serializeString(pkg.Name, s) //Change Name to String
 }
 
-func serializeStructName(strct *CXStruct, s *serializedCXProgram) {
+func serializeStructName(strct *CXStruct, s *SerializedCXProgram) {
 	strctName := strct.Package.Name + "." + strct.Name
 	sStrct := &s.Structs[s.StructsMap[strctName]]
 	sStrct.NameOffset, sStrct.NameSize = serializeString(strct.Name, s) //Change Name to String
 }
 
-func serializeFunctionName(fn *CXFunction, s *serializedCXProgram) {
+func serializeFunctionName(fn *CXFunction, s *SerializedCXProgram) {
 	fnName := fn.Package.Name + "." + fn.Name
 	if off, found := s.FunctionsMap[fnName]; found {
 		sFn := &s.Functions[off]
@@ -481,7 +273,7 @@ func serializeFunctionName(fn *CXFunction, s *serializedCXProgram) {
 	}
 }
 
-func serializePackageGlobals(pkg *CXPackage, s *serializedCXProgram) {
+func serializePackageGlobals(pkg *CXPackage, s *SerializedCXProgram) {
 	if pkgOff, found := s.PackagesMap[pkg.Name]; found {
 		sPkg := &s.Packages[pkgOff]
 		sPkg.GlobalsOffset, sPkg.GlobalsSize = serializeSliceOfArguments(pkg.Globals, s)
@@ -490,7 +282,7 @@ func serializePackageGlobals(pkg *CXPackage, s *serializedCXProgram) {
 	}
 }
 
-func serializePackageImports(pkg *CXPackage, s *serializedCXProgram) {
+func serializePackageImports(pkg *CXPackage, s *SerializedCXProgram) {
 	l := len(pkg.Imports)
 	if l == 0 {
 		s.Packages[s.PackagesMap[pkg.Name]].ImportsOffset = int64(-1)
@@ -511,7 +303,7 @@ func serializePackageImports(pkg *CXPackage, s *serializedCXProgram) {
 	s.Integers = append(s.Integers, imps...)
 }
 
-func serializeStructPackage(strct *CXStruct, s *serializedCXProgram) {
+func serializeStructPackage(strct *CXStruct, s *SerializedCXProgram) {
 	strctName := strct.Package.Name + "." + strct.Name
 	if pkgOff, found := s.PackagesMap[strct.Package.Name]; found {
 		if off, found := s.StructsMap[strctName]; found {
@@ -525,7 +317,7 @@ func serializeStructPackage(strct *CXStruct, s *serializedCXProgram) {
 	}
 }
 
-func serializeFunctionPackage(fn *CXFunction, s *serializedCXProgram) {
+func serializeFunctionPackage(fn *CXFunction, s *SerializedCXProgram) {
 	fnName := fn.Package.Name + "." + fn.Name
 	if pkgOff, found := s.PackagesMap[fn.Package.Name]; found {
 		if off, found := s.FunctionsMap[fnName]; found {
@@ -539,7 +331,7 @@ func serializeFunctionPackage(fn *CXFunction, s *serializedCXProgram) {
 	}
 }
 
-func serializePackageIntegers(pkg *CXPackage, s *serializedCXProgram) {
+func serializePackageIntegers(pkg *CXPackage, s *SerializedCXProgram) {
 	if pkgOff, found := s.PackagesMap[pkg.Name]; found {
 		sPkg := &s.Packages[pkgOff]
 
@@ -573,7 +365,7 @@ func serializePackageIntegers(pkg *CXPackage, s *serializedCXProgram) {
 	}
 }
 
-func serializeStructIntegers(strct *CXStruct, s *serializedCXProgram) {
+func serializeStructIntegers(strct *CXStruct, s *SerializedCXProgram) {
 	strctName := strct.Package.Name + "." + strct.Name
 	if off, found := s.StructsMap[strctName]; found {
 		sStrct := &s.Structs[off]
@@ -583,7 +375,7 @@ func serializeStructIntegers(strct *CXStruct, s *serializedCXProgram) {
 	}
 }
 
-func serializeFunctionIntegers(fn *CXFunction, s *serializedCXProgram) {
+func serializeFunctionIntegers(fn *CXFunction, s *SerializedCXProgram) {
 	fnName := fn.Package.Name + "." + fn.Name
 	if off, found := s.FunctionsMap[fnName]; found {
 		sFn := &s.Functions[off]
@@ -598,11 +390,11 @@ func serializeFunctionIntegers(fn *CXFunction, s *serializedCXProgram) {
 // container for our serialized cx program.
 // Program memory is also added here to our container
 // if memory is to be included.
-func initSerialization(prgrm *CXProgram, s *serializedCXProgram, includeMemory bool) {
-	s.PackagesMap = make(map[string]int)
-	s.StructsMap = make(map[string]int)
-	s.FunctionsMap = make(map[string]int)
-	s.StringsMap = make(map[string]int)
+func initSerialization(prgrm *CXProgram, s *SerializedCXProgram, includeMemory bool) {
+	s.PackagesMap = make(map[string]int64)
+	s.StructsMap = make(map[string]int64)
+	s.FunctionsMap = make(map[string]int64)
+	s.StringsMap = make(map[string]int64)
 
 	s.Calls = make([]serializedCall, prgrm.CallCounter)
 	s.Packages = make([]serializedPackage, len(prgrm.Packages))
@@ -627,7 +419,7 @@ func initSerialization(prgrm *CXProgram, s *serializedCXProgram, includeMemory b
 
 // serializeProgram serializes
 // program of cx program.
-func serializeProgram(prgrm *CXProgram, s *serializedCXProgram) {
+func serializeProgram(prgrm *CXProgram, s *SerializedCXProgram) {
 	s.Program = serializedProgram{}
 	sPrgrm := &s.Program
 	sPrgrm.PackagesOffset = int64(0)
@@ -663,7 +455,7 @@ func serializeProgram(prgrm *CXProgram, s *serializedCXProgram) {
 
 // serializeCXProgramElements is used serializing CX program's
 // elements (packages, structs, functions, etc.).
-func serializeCXProgramElements(prgrm *CXProgram, s *serializedCXProgram) {
+func serializeCXProgramElements(prgrm *CXProgram, s *SerializedCXProgram) {
 	var fnCounter int64
 	var strctCounter int64
 
@@ -781,7 +573,7 @@ func serializeCXProgramElements(prgrm *CXProgram, s *serializedCXProgram) {
 // These slice of bytes can then be deserialize in the future and
 // be translated back to cx program.
 func SerializeCXProgram(prgrm *CXProgram, includeMemory bool) (b []byte) {
-	s := serializedCXProgram{}
+	s := SerializedCXProgram{}
 	initSerialization(prgrm, &s, includeMemory)
 
 	// serialize cx program's packages,
@@ -836,7 +628,7 @@ func SerializeCXProgram(prgrm *CXProgram, includeMemory bool) (b []byte) {
 // SerializeDebugInfo prints the name of the serialized segment and byte size.
 func SerializeDebugInfo(prgrm *CXProgram, includeMemory bool) SerializedDataSize {
 	idxSize := encoder.Size(serializedCXProgramIndex{})
-	var s serializedCXProgram
+	var s SerializedCXProgram
 
 	bytes := SerializeCXProgram(prgrm, includeMemory)
 	helper.DeserializeRaw(bytes[:idxSize], &s.Index)
@@ -857,7 +649,7 @@ func SerializeDebugInfo(prgrm *CXProgram, includeMemory bool) SerializedDataSize
 	return *data
 }
 
-func deserializeString(off int64, size int64, s *serializedCXProgram) string {
+func deserializeString(off int64, size int64, s *SerializedCXProgram) string {
 	if size < 1 {
 		return ""
 	}
@@ -868,7 +660,7 @@ func deserializeString(off int64, size int64, s *serializedCXProgram) string {
 	return name
 }
 
-func deserializePackages(s *serializedCXProgram, prgrm *CXProgram) {
+func deserializePackages(s *SerializedCXProgram, prgrm *CXProgram) {
 	var fnCounter int64
 	var strctCounter int64
 
@@ -964,14 +756,14 @@ func deserializePackages(s *serializedCXProgram, prgrm *CXProgram) {
 	prgrm.CurrentPackage = prgrm.Packages[s.Program.CurrentPackageOffset]
 }
 
-func deserializeStruct(sStrct *serializedStruct, strct *CXStruct, s *serializedCXProgram, prgrm *CXProgram) {
+func deserializeStruct(sStrct *serializedStruct, strct *CXStruct, s *SerializedCXProgram, prgrm *CXProgram) {
 	strct.Name = deserializeString(sStrct.NameOffset, sStrct.NameSize, s)
 	strct.Fields = deserializeArguments(sStrct.FieldsOffset, sStrct.FieldsSize, s, prgrm)
 	strct.Size = int(sStrct.Size)
 	strct.Package = prgrm.Packages[sStrct.PackageOffset]
 }
 
-func deserializeArguments(off int64, size int64, s *serializedCXProgram, prgrm *CXProgram) []*CXArgument {
+func deserializeArguments(off int64, size int64, s *SerializedCXProgram, prgrm *CXProgram) []*CXArgument {
 	if size < 1 {
 		return nil
 	}
@@ -987,7 +779,7 @@ func deserializeArguments(off int64, size int64, s *serializedCXProgram, prgrm *
 	return args
 }
 
-// func getStructType(sArg *serializedArgument, s *serializedCXProgram, prgrm *CXProgram) *CXStruct {
+// func getStructType(sArg *serializedArgument, s *SerializedCXProgram, prgrm *CXProgram) *CXStruct {
 // 	if sArg.StructTypeOffset < 0 {
 // 		return nil
 // 	}
@@ -1005,7 +797,7 @@ func deserializeArguments(off int64, size int64, s *serializedCXProgram, prgrm *
 // 	return nil
 // }
 
-func deserializeArgument(sArg *serializedArgument, s *serializedCXProgram, prgrm *CXProgram) *CXArgument {
+func deserializeArgument(sArg *serializedArgument, s *SerializedCXProgram, prgrm *CXProgram) *CXArgument {
 	var arg CXArgument
 	arg.ArgDetails.Name = deserializeString(sArg.NameOffset, sArg.NameSize, s)
 	arg.Type = int(sArg.Type)
@@ -1046,7 +838,7 @@ func deserializeArgument(sArg *serializedArgument, s *serializedCXProgram, prgrm
 	return &arg
 }
 
-func deserializeOperator(sExpr *serializedExpression, s *serializedCXProgram, prgrm *CXProgram) *CXFunction {
+func deserializeOperator(sExpr *serializedExpression, s *SerializedCXProgram, prgrm *CXProgram) *CXFunction {
 	if sExpr.OperatorOffset < 0 {
 		return nil
 	}
@@ -1064,7 +856,7 @@ func deserializeOperator(sExpr *serializedExpression, s *serializedCXProgram, pr
 	return nil
 }
 
-func deserializePackageImport(sImp *serializedPackage, s *serializedCXProgram, prgrm *CXProgram) *CXPackage {
+func deserializePackageImport(sImp *serializedPackage, s *SerializedCXProgram, prgrm *CXProgram) *CXPackage {
 	impName := deserializeString(sImp.NameOffset, sImp.NameSize, s)
 
 	for _, pkg := range prgrm.Packages {
@@ -1076,7 +868,7 @@ func deserializePackageImport(sImp *serializedPackage, s *serializedCXProgram, p
 	return nil
 }
 
-func deserializeExpressionFunction(sExpr *serializedExpression, s *serializedCXProgram, prgrm *CXProgram) *CXFunction {
+func deserializeExpressionFunction(sExpr *serializedExpression, s *SerializedCXProgram, prgrm *CXProgram) *CXFunction {
 	if sExpr.FunctionOffset < 0 {
 		return nil
 	}
@@ -1094,7 +886,7 @@ func deserializeExpressionFunction(sExpr *serializedExpression, s *serializedCXP
 	return nil
 }
 
-func deserializeExpressions(off int64, size int64, s *serializedCXProgram, prgrm *CXProgram) []*CXExpression {
+func deserializeExpressions(off int64, size int64, s *SerializedCXProgram, prgrm *CXProgram) []*CXExpression {
 	if size < 1 {
 		return nil
 	}
@@ -1110,7 +902,7 @@ func deserializeExpressions(off int64, size int64, s *serializedCXProgram, prgrm
 	return exprs
 }
 
-func deserializeExpression(sExpr *serializedExpression, s *serializedCXProgram, prgrm *CXProgram) *CXExpression {
+func deserializeExpression(sExpr *serializedExpression, s *SerializedCXProgram, prgrm *CXProgram) *CXExpression {
 	var expr CXExpression
 
 	if deserializeBool(sExpr.IsNative) {
@@ -1135,7 +927,7 @@ func deserializeExpression(sExpr *serializedExpression, s *serializedCXProgram, 
 	return &expr
 }
 
-func deserializeFunction(sFn *serializedFunction, fn *CXFunction, s *serializedCXProgram, prgrm *CXProgram) {
+func deserializeFunction(sFn *serializedFunction, fn *CXFunction, s *SerializedCXProgram, prgrm *CXProgram) {
 	fn.Name = deserializeString(sFn.NameOffset, sFn.NameSize, s)
 	fn.Inputs = deserializeArguments(sFn.InputsOffset, sFn.InputsSize, s, prgrm)
 	fn.Outputs = deserializeArguments(sFn.OutputsOffset, sFn.OutputsSize, s, prgrm)
@@ -1155,7 +947,7 @@ func deserializeBool(val int64) bool {
 	return val == 1
 }
 
-func deserializeIntegers(off int64, size int64, s *serializedCXProgram) []int {
+func deserializeIntegers(off int64, size int64, s *SerializedCXProgram) []int {
 	if size < 1 {
 		return nil
 	}
@@ -1169,7 +961,7 @@ func deserializeIntegers(off int64, size int64, s *serializedCXProgram) []int {
 }
 
 // initDeserialization initializes the CXProgram fields that represent a CX program. This should be refactored, as the names Deserialize and initDeserialization create some naming conflict.
-func initDeserialization(prgrm *CXProgram, s *serializedCXProgram) {
+func initDeserialization(prgrm *CXProgram, s *SerializedCXProgram) {
 	prgrm.Memory = s.Memory
 	prgrm.Packages = make([]*CXPackage, len(s.Packages))
 	prgrm.CallStack = make([]CXCall, constants.CALLSTACK_SIZE)
@@ -1189,7 +981,7 @@ func Deserialize(b []byte) (prgrm *CXProgram) {
 	prgrm = &CXProgram{}
 	idxSize := encoder.Size(serializedCXProgramIndex{})
 
-	var s serializedCXProgram
+	var s SerializedCXProgram
 
 	helper.DeserializeRaw(b[:idxSize], &s.Index)
 	helper.DeserializeRaw(b[s.Index.ProgramOffset:s.Index.CallsOffset], &s.Program)
